backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pillow==10.2.0
numpy==1.26.3
opencv-python==4.9.0.80
torch==2.1.2
torchvision==0.16.2
diffusers==0.25.1
transformers==4.37.2
accelerate==0.26.1
rembg==2.0.56
mediapipe==0.10.9
scipy==1.11.4
pydantic==2.5.3
python-dotenv==1.0.0
aiofiles==23.2.1
